pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discoursemap"
version = "2.1.0"
description = "Discourse forum security scanner. Written for security professionals and forum administrators."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ibrahimsql", email = "ibrahimsql@proton.me"}
]
maintainers = [
    {name = "ibrahimsql", email = "ibrahimsql@proton.me"}
]
keywords = ["discourse", "security", "scanner", "penetration-testing", "vulnerability-scanner", "cybersecurity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
    "PyYAML==6.0.3",
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "lxml>=6.0.0",
    "colorama>=0.4.6",
    "tqdm>=4.67.0",
    "jinja2>=3.1.6",
    "urllib3>=2.5.0",
    "certifi>=2025.7.14",
    "chardet>=5.2.0",
    "idna>=3.10",
    "pysocks>=1.7.1",
    "cryptography>=45.0.0",
    "pyopenssl>=25.1.0",
    "typing-extensions>=4.0.0",
]



# Test and dev dependencies removed since tests folder is empty
# [project.optional-dependencies]
# test = [
#     "pytest>=7.0.0",
#     "pytest-cov>=4.0.0",
#     "pytest-mock>=3.10.0",
# ]
# dev = [
#     "pytest>=7.0.0",
#     "pytest-cov>=4.0.0",
#     "pytest-mock>=3.10.0",
#     "black>=23.0.0",
#     "flake8>=6.0.0",
#     "mypy>=1.0.0",
# ]

[project.urls]
"Homepage" = "https://github.com/ibrahmsql/discoursemap"
"Bug Reports" = "https://github.com/ibrahmsql/discoursemap/issues"
"Source" = "https://github.com/ibrahmsql/discoursemap"
"Documentation" = "https://github.com/ibrahmsql/discoursemap#readme"

[project.scripts]
discoursemap = "discoursemap.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."] 
include = ["discoursemap*"]

[tool.setuptools.package-data]
discoursemap = [
    "data/*.yaml",
]